refactored datastore mongodb impl
