+++ data-access/src/app/external-dependencies/domain.ts
@@ -1,40 +1,40 @@
 // community
-export { MongoCommunityUnitOfWork as CommunityUnitOfWork} from '../../infrastructure-services-impl/datastore/mongodb/infrastructure/community.mongo-uow';
-export { CommunityDomainAdapter, CommunityConverter } from '../../infrastructure-services-impl/datastore/mongodb/infrastructure/community.domain-adapter';
-export { MongoCommunityRepository as CommunityRepository } from '../../infrastructure-services-impl/datastore/mongodb/infrastructure/community.mongo-repository';
+export { MongoCommunityUnitOfWork as CommunityUnitOfWork} from '../../infrastructure-services-impl/datastore/mongodb/infrastructure/community/community.mongo-uow';
+export { CommunityDomainAdapter, CommunityConverter } from '../../infrastructure-services-impl/datastore/mongodb/infrastructure/community/community.domain-adapter';
+export { MongoCommunityRepository as CommunityRepository } from '../../infrastructure-services-impl/datastore/mongodb/infrastructure/community/community.mongo-repository';
 
 // member
-export { MongoMemberUnitOfWork as MemberUnitOfWork } from '../../infrastructure-services-impl/datastore/mongodb/infrastructure/member.mongo-uow';
-export { MemberDomainAdapter, MemberConverter } from '../../infrastructure-services-impl/datastore/mongodb/infrastructure/member.domain-adapter';
-export { MongoMemberRepository as MemberRepository } from '../../infrastructure-services-impl/datastore/mongodb/infrastructure/member.mongo-repository';
+export { MongoMemberUnitOfWork as MemberUnitOfWork } from '../../infrastructure-services-impl/datastore/mongodb/infrastructure/member/member.mongo-uow';
+export { MemberDomainAdapter, MemberConverter } from '../../infrastructure-services-impl/datastore/mongodb/infrastructure/member/member.domain-adapter';
+export { MongoMemberRepository as MemberRepository } from '../../infrastructure-services-impl/datastore/mongodb/infrastructure/member/member.mongo-repository';
 
 // role
-export { MongoRoleUnitOfWork as RoleUnitOfWork } from '../../infrastructure-services-impl/datastore/mongodb/infrastructure/role.mongo-uow';
-export { RoleDomainAdapter, RoleConverter } from '../../infrastructure-services-impl/datastore/mongodb/infrastructure/role.domain-adapter';
-export { MongoRoleRepository as RoleRepository } from '../../infrastructure-services-impl/datastore/mongodb/infrastructure/role.mongo-repository';
+export { MongoRoleUnitOfWork as RoleUnitOfWork } from '../../infrastructure-services-impl/datastore/mongodb/infrastructure/role/role.mongo-uow';
+export { RoleDomainAdapter, RoleConverter } from '../../infrastructure-services-impl/datastore/mongodb/infrastructure/role/role.domain-adapter';
+export { MongoRoleRepository as RoleRepository } from '../../infrastructure-services-impl/datastore/mongodb/infrastructure/role/role.mongo-repository';
 
 // property
-export { MongoPropertyUnitOfWork as PropertyUnitOfWork } from '../../infrastructure-services-impl/datastore/mongodb/infrastructure/property.mongo-uow';
-export { PropertyDomainAdapter, PropertyConverter } from '../../infrastructure-services-impl/datastore/mongodb/infrastructure/property.domain-adapter';
-export { MongoPropertyRepository as PropertyRepository } from '../../infrastructure-services-impl/datastore/mongodb/infrastructure/property.mongo-repository';
+export { MongoPropertyUnitOfWork as PropertyUnitOfWork } from '../../infrastructure-services-impl/datastore/mongodb/infrastructure/property/property.mongo-uow';
+export { PropertyDomainAdapter, PropertyConverter } from '../../infrastructure-services-impl/datastore/mongodb/infrastructure/property/property.domain-adapter';
+export { MongoPropertyRepository as PropertyRepository } from '../../infrastructure-services-impl/datastore/mongodb/infrastructure/property/property.mongo-repository';
 
 // service
-export { MongoServiceUnitOfWork as ServiceUnitOfWork } from '../../infrastructure-services-impl/datastore/mongodb/infrastructure/service.uow';
-export { ServiceDomainAdapter, ServiceConverter } from '../../infrastructure-services-impl/datastore/mongodb/infrastructure/service.domain-adapter';
-export { MongoServiceRepository as ServiceRepository } from '../../infrastructure-services-impl/datastore/mongodb/infrastructure/service.mongo-repository';
+export { MongoServiceUnitOfWork as ServiceUnitOfWork } from '../../infrastructure-services-impl/datastore/mongodb/infrastructure/service/service.uow';
+export { ServiceDomainAdapter, ServiceConverter } from '../../infrastructure-services-impl/datastore/mongodb/infrastructure/service/service.domain-adapter';
+export { MongoServiceRepository as ServiceRepository } from '../../infrastructure-services-impl/datastore/mongodb/infrastructure/service/service.mongo-repository';
 
 // service-ticket
-export { MongoServiceTicketUnitOfWork as ServiceTicketUnitOfWork } from '../../infrastructure-services-impl/datastore/mongodb/infrastructure/service-ticket.uow';
-export { ServiceTicketDomainAdapter, ServiceTicketConverter } from '../../infrastructure-services-impl/datastore/mongodb/infrastructure/service-ticket.domain-adapter';
-export { MongoServiceTicketRepository as ServiceTicketRepository } from '../../infrastructure-services-impl/datastore/mongodb/infrastructure/service-ticket.mongo-repository';
+export { MongoServiceTicketUnitOfWork as ServiceTicketUnitOfWork } from '../../infrastructure-services-impl/datastore/mongodb/infrastructure/service-ticket/v1/service-ticket.uow';
+export { ServiceTicketDomainAdapter, ServiceTicketConverter } from '../../infrastructure-services-impl/datastore/mongodb/infrastructure/service-ticket/v1/service-ticket.domain-adapter';
+export { MongoServiceTicketRepository as ServiceTicketRepository } from '../../infrastructure-services-impl/datastore/mongodb/infrastructure/service-ticket/v1/service-ticket.mongo-repository';
 
 // violation-ticket
-export { MongoViolationTicketUnitOfWork as ViolationTicketUnitOfWork } from '../../infrastructure-services-impl/datastore/mongodb/infrastructure/violation-ticket.uow';
-export { ViolationTicketDomainAdapter as ViolationTicketDomainAdapter, ViolationTicketConverter } from '../../infrastructure-services-impl/datastore/mongodb/infrastructure/violation-ticket.domain-adapter';
-export { MongoViolationTicketRepository as ViolationTicketRepository } from '../../infrastructure-services-impl/datastore/mongodb/infrastructure/violation-ticket.mongo-repository';
+export { MongoViolationTicketUnitOfWork as ViolationTicketUnitOfWork } from '../../infrastructure-services-impl/datastore/mongodb/infrastructure/violation-ticket/v1/violation-ticket.uow';
+export { ViolationTicketDomainAdapter as ViolationTicketDomainAdapter, ViolationTicketConverter } from '../../infrastructure-services-impl/datastore/mongodb/infrastructure/violation-ticket/v1/violation-ticket.domain-adapter';
+export { MongoViolationTicketRepository as ViolationTicketRepository } from '../../infrastructure-services-impl/datastore/mongodb/infrastructure/violation-ticket/v1/violation-ticket.mongo-repository';
 
 // user
-export { MongoUserUnitOfWork as UserUnitOfWork } from '../../infrastructure-services-impl/datastore/mongodb/infrastructure/user.uow';
-export { UserDomainAdapter, UserConverter } from '../../infrastructure-services-impl/datastore/mongodb/infrastructure/user.domain-adapter';
-export { MongoUserRepository as UserRepository } from '../../infrastructure-services-impl/datastore/mongodb/infrastructure/user.mongo-repository';
+export { MongoUserUnitOfWork as UserUnitOfWork } from '../../infrastructure-services-impl/datastore/mongodb/infrastructure/user/user.uow';
+export { UserDomainAdapter, UserConverter } from '../../infrastructure-services-impl/datastore/mongodb/infrastructure/user/user.domain-adapter';
+export { MongoUserRepository as UserRepository } from '../../infrastructure-services-impl/datastore/mongodb/infrastructure/user/user.mongo-repository';
