+++ data-access/src/infrastructure-services-impl/datastore/mongodb/infrastructure/community/community.domain-adapter.ts
@@ -1,10 +1,10 @@
-import { Community } from '../models/community';
-import { Community as CommunityDO, CommunityProps } from '../../../../app/domain/contexts/community/community/community';
-import { MongooseDomainAdapter } from '../../../../../seedwork/services-seedwork-datastore-mongodb/infrastructure/mongo-domain-adapter';
-import { MongoTypeConverter } from '../../../../../seedwork/services-seedwork-datastore-mongodb/infrastructure/mongo-type-converter';
-import { DomainExecutionContext } from '../../../../app/domain/contexts/domain-execution-context';
-import { UserEntityReference, UserProps } from '../../../../app/domain/contexts/user/user';
-import { UserDomainAdapter } from './user.domain-adapter';
+import { Community } from '../../models/community';
+import { Community as CommunityDO, CommunityProps } from '../../../../../app/domain/contexts/community/community/community';
+import { MongooseDomainAdapter } from '../../../../../../seedwork/services-seedwork-datastore-mongodb/infrastructure/mongo-domain-adapter';
+import { MongoTypeConverter } from '../../../../../../seedwork/services-seedwork-datastore-mongodb/infrastructure/mongo-type-converter';
+import { DomainExecutionContext } from '../../../../../app/domain/contexts/domain-execution-context';
+import { UserEntityReference, UserProps } from '../../../../../app/domain/contexts/user/user';
+import { UserDomainAdapter } from '../user/user.domain-adapter';
 
 export class CommunityConverter extends MongoTypeConverter<DomainExecutionContext, Community, CommunityDomainAdapter, CommunityDO<CommunityDomainAdapter>> {
   constructor() {
