--- data-access/src/infrastructure-services-impl/datastore/mongodb/infrastructure/member.mongo-uow.ts
@@ -1,7 +1,0 @@
-import { MongoUnitOfWork } from '../../../../../seedwork/services-seedwork-datastore-mongodb/infrastructure/mongo-unit-of-work';
-import { MemberModel } from '../models/member';
-import { MemberConverter } from './member.domain-adapter';
-import { MongoMemberRepository } from './member.mongo-repository';
-import { InProcEventBusInstance, NodeEventBusInstance } from '../../../../../seedwork/event-bus-seedwork-node';
-
-export const MongoMemberUnitOfWork = new MongoUnitOfWork(InProcEventBusInstance, NodeEventBusInstance, MemberModel, new MemberConverter(), MongoMemberRepository);
