+++ data-access/src/infrastructure-services-impl/datastore/mongodb/infrastructure/property/property.mongo-repository.ts
@@ -1,9 +1,9 @@
-import { Property as PropertyDO, PropertyProps } from '../../../../app/domain/contexts/property/property';
-import { PropertyRepository } from '../../../../app/domain/contexts/property/property.repository';
-import { Property } from '../models/property';
-import { MongoRepositoryBase } from '../../../../../seedwork/services-seedwork-datastore-mongodb/infrastructure/mongo-repository';
-import { DomainExecutionContext } from '../../../../app/domain/contexts/domain-execution-context';
-import { CommunityEntityReference } from '../../../../app/domain/contexts/community/community/community';
+import { Property as PropertyDO, PropertyProps } from '../../../../../app/domain/contexts/property/property';
+import { PropertyRepository } from '../../../../../app/domain/contexts/property/property.repository';
+import { Property } from '../../models/property';
+import { MongoRepositoryBase } from '../../../../../../seedwork/services-seedwork-datastore-mongodb/infrastructure/mongo-repository';
+import { DomainExecutionContext } from '../../../../../app/domain/contexts/domain-execution-context';
+import { CommunityEntityReference } from '../../../../../app/domain/contexts/community/community/community';
 
 export class MongoPropertyRepository<PropType extends PropertyProps>
   extends MongoRepositoryBase<DomainExecutionContext, Property, PropType, PropertyDO<PropType>>
