+++ data-access/src/infrastructure-services-impl/datastore/mongodb/infrastructure/property/property.mongo-uow.ts
@@ -1,7 +1,7 @@
-import { MongoUnitOfWork } from '../../../../../seedwork/services-seedwork-datastore-mongodb/infrastructure/mongo-unit-of-work';
-import { PropertyModel } from '../models/property';
+import { MongoUnitOfWork } from '../../../../../../seedwork/services-seedwork-datastore-mongodb/infrastructure/mongo-unit-of-work';
+import { PropertyModel } from '../../models/property';
 import { PropertyConverter } from './property.domain-adapter';
 import { MongoPropertyRepository } from './property.mongo-repository';
-import { InProcEventBusInstance, NodeEventBusInstance } from '../../../../../seedwork/event-bus-seedwork-node';
+import { InProcEventBusInstance, NodeEventBusInstance } from '../../../../../../seedwork/event-bus-seedwork-node';
 
 export const MongoPropertyUnitOfWork = new MongoUnitOfWork(InProcEventBusInstance, NodeEventBusInstance, PropertyModel, new PropertyConverter(), MongoPropertyRepository);
