+++ data-access/src/infrastructure-services-impl/datastore/mongodb/infrastructure/role/role.mongo-repository.ts
@@ -1,9 +1,9 @@
-import { Role as RoleDO, RoleProps } from '../../../../app/domain/contexts/community/role/role';
-import { RoleRepository } from '../../../../app/domain/contexts/community/role/role.repository';
-import { Role } from '../models/role';
-import { MongoRepositoryBase } from '../../../../../seedwork/services-seedwork-datastore-mongodb/infrastructure/mongo-repository';
-import { DomainExecutionContext } from '../../../../app/domain/contexts/domain-execution-context';
-import { CommunityEntityReference } from '../../../../app/domain/contexts/community/community/community';
+import { Role as RoleDO, RoleProps } from '../../../../../app/domain/contexts/community/role/role';
+import { RoleRepository } from '../../../../../app/domain/contexts/community/role/role.repository';
+import { Role } from '../../models/role';
+import { MongoRepositoryBase } from '../../../../../../seedwork/services-seedwork-datastore-mongodb/infrastructure/mongo-repository';
+import { DomainExecutionContext } from '../../../../../app/domain/contexts/domain-execution-context';
+import { CommunityEntityReference } from '../../../../../app/domain/contexts/community/community/community';
 
 export class MongoRoleRepository<PropType extends RoleProps>
   extends MongoRepositoryBase<DomainExecutionContext, Role, PropType, RoleDO<PropType>>
