+++ data-access/src/infrastructure-services-impl/datastore/mongodb/infrastructure/service/service.domain-adapter.ts
@@ -1,11 +1,11 @@
-import { Service } from '../models/service';
-import { Service as ServiceDO, ServiceProps } from '../../../../app/domain/contexts/community/service/service';
-import { MongooseDomainAdapter } from '../../../../../seedwork/services-seedwork-datastore-mongodb/infrastructure/mongo-domain-adapter';
-import { MongoTypeConverter } from '../../../../../seedwork/services-seedwork-datastore-mongodb/infrastructure/mongo-type-converter';
+import { Service } from '../../models/service';
+import { Service as ServiceDO, ServiceProps } from '../../../../../app/domain/contexts/community/service/service';
+import { MongooseDomainAdapter } from '../../../../../../seedwork/services-seedwork-datastore-mongodb/infrastructure/mongo-domain-adapter';
+import { MongoTypeConverter } from '../../../../../../seedwork/services-seedwork-datastore-mongodb/infrastructure/mongo-type-converter';
 
-import { CommunityProps } from '../../../../app/domain/contexts/community/community/community';
-import { CommunityDomainAdapter } from './community.domain-adapter';
-import { DomainExecutionContext } from '../../../../app/domain/contexts/domain-execution-context';
+import { CommunityProps } from '../../../../../app/domain/contexts/community/community/community';
+import { CommunityDomainAdapter } from '../community/community.domain-adapter';
+import { DomainExecutionContext } from '../../../../../app/domain/contexts/domain-execution-context';
 
 export class ServiceConverter extends MongoTypeConverter<DomainExecutionContext, Service, ServiceDomainAdapter, ServiceDO<ServiceDomainAdapter>> {
   constructor() {
