+++ data-access/src/infrastructure-services-impl/datastore/mongodb/infrastructure/user/user.mongo-repository.ts
@@ -1,8 +1,8 @@
-import { User as UserDO, UserProps } from '../../../../app/domain/contexts/user/user';
-import { UserRepository } from '../../../../app/domain/contexts/user/user.repository';
-import { User } from '../models/user';
-import { MongoRepositoryBase } from '../../../../../seedwork/services-seedwork-datastore-mongodb/infrastructure/mongo-repository';
-import { DomainExecutionContext } from '../../../../app/domain/contexts/domain-execution-context';
+import { User as UserDO, UserProps } from '../../../../../app/domain/contexts/user/user';
+import { UserRepository } from '../../../../../app/domain/contexts/user/user.repository';
+import { User } from '../../models/user';
+import { MongoRepositoryBase } from '../../../../../../seedwork/services-seedwork-datastore-mongodb/infrastructure/mongo-repository';
+import { DomainExecutionContext } from '../../../../../app/domain/contexts/domain-execution-context';
 
 export class MongoUserRepository<PropType extends UserProps>
   extends MongoRepositoryBase<DomainExecutionContext, User, PropType, UserDO<PropType>>
