+++ data-access/src/infrastructure-services-impl/datastore/mongodb/infrastructure/violation-ticket/v1/violation-ticket.uow.ts
@@ -1,7 +1,7 @@
 
-import { MongoUnitOfWork } from '../../../../../seedwork/services-seedwork-datastore-mongodb/infrastructure/mongo-unit-of-work';
-import { ViolationTicketModel } from '../models/violation-ticket';
-import { InProcEventBusInstance, NodeEventBusInstance } from '../../../../../seedwork/event-bus-seedwork-node';
+import { MongoUnitOfWork } from '../../../../../../../seedwork/services-seedwork-datastore-mongodb/infrastructure/mongo-unit-of-work';
+import { ViolationTicketModel } from '../../../models/violation-ticket';
+import { InProcEventBusInstance, NodeEventBusInstance } from '../../../../../../../seedwork/event-bus-seedwork-node';
 import { ViolationTicketConverter } from './violation-ticket.domain-adapter';
 import { MongoViolationTicketRepository as ViolationTicketRepository } from './violation-ticket.mongo-repository';
